backend/src/audiorating_backend/models.py
from datetime import datetime
from typing import Optional, List, Dict, Any
from sqlmodel import SQLModel, Field, Column, Relationship
from sqlalchemy.dialects.postgresql import JSONB
from pydantic import model_validator, ConfigDict
import uuid

